refactor(layering): AppError via appError.js (47 services), prospectScope + SGT maths out of wrong layers (P4-7/3)
