+++ backend/src/services/authService.js
@@ -2,7 +2,7 @@ import { Op } from 'sequelize';
 import { v4 as uuidv4 } from 'uuid';
 import { User } from '../models/index.js';
 import { generateToken } from '../middleware/auth.js';
-import { AppError } from '../middleware/errorHandler.js';
+import { AppError } from '../middleware/appError.js';
 import { logger } from '../utils/logger.js';
 import { isAllowedPublicHost } from '../utils/publicHost.js';
