+++ backend/src/services/campaignDrawGuards.js
@@ -1,6 +1,6 @@
 import crypto from 'crypto';
 import { DrawTermsVersion, Draw } from '../models/index.js';
-import { AppError } from '../middleware/errorHandler.js';
+import { AppError } from '../middleware/appError.js';
 import { logger } from '../utils/logger.js';
 import { sgtDayEndExclusiveMs } from '../utils/sgtTime.js';
 import { normalizeLuckyDraw, assertSingleWinnerDraw } from '../utils/luckyDraw.js';
