+++ backend/src/services/cohortService.js
@@ -1,5 +1,5 @@
 import { sequelize, Cohort, Campaign } from '../models/index.js';
-import { AppError } from '../middleware/errorHandler.js';
+import { AppError } from '../middleware/appError.js';
 import { logger } from '../utils/logger.js';
 import { CONSUMER_CATEGORIES, CONSUMER_CATEGORY_DEFS } from '../utils/marketplaceContent.js';
 import { getStoredCategory } from '../utils/designConfigV2Clamp.js';
