+++ backend/src/services/guidedReviewAiService.js
@@ -1,4 +1,4 @@
-import { AppError } from '../middleware/errorHandler.js';
+import { AppError } from '../middleware/appError.js';
 import { logger } from '../utils/logger.js';
 
 const stringField = { type: 'string', minLength: 1, maxLength: 500 };
