+++ backend/src/services/leadPackageService.js
@@ -1,6 +1,6 @@
 import { Op } from 'sequelize';
 import { LeadPackage, LeadPackageAssignment, User, Campaign, Prospect, sequelize } from '../models/index.js';
-import { AppError } from '../middleware/errorHandler.js';
+import { AppError } from '../middleware/appError.js';
 import { logger } from '../utils/logger.js';
 
 /**
