+++ backend/src/services/redeemOps/cadenceService.js
@@ -4,7 +4,7 @@ import {
   OutreachSuppression, OutreachTask, PartnerOrganisation, PartnerContact, PartnerLocation,
   User, sequelize,
 } from '../../models/index.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 import { hasCapability, canActOnPartnerRow, isManagerTier } from './permissions.js';
 import { makeRedeemOpsAuditService } from './auditService.js';
