+++ backend/src/services/redeemOps/discoveryAiService.js
@@ -1,4 +1,4 @@
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 import { cfg } from './discoveryService.js';
 import { staffFacingAiError } from './aiSuggestShared.js';
