+++ backend/src/services/redeemOps/entitlementService.js
@@ -4,7 +4,7 @@ import {
   PartnerOrganisation, Prospect, User, Consumer, Campaign, WaMessageStatus, sequelize,
 } from '../../models/index.js';
 import { phoneVerificationIsCurrent } from '../consumerService.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 import { makeInventoryService } from './inventoryService.js';
 import { makeRedeemOpsAuditService } from './auditService.js';
