+++ backend/src/services/redeemOps/inventoryService.js
@@ -1,5 +1,5 @@
 import { RewardOffer, RewardInventoryEvent, sequelize } from '../../models/index.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 
 /**
