+++ backend/src/services/redeemOps/onboardingService.js
@@ -1,5 +1,5 @@
 import { PartnerOnboardingItem, PartnerOrganisation, sequelize } from '../../models/index.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 
 /**
