+++ backend/src/services/redeemOps/partnerService.js
@@ -6,7 +6,7 @@ import {
   RewardEntitlement, Redemption, RewardInventoryEvent, Draw,
   RedeemOpsAuditEvent, User, sequelize,
 } from '../../models/index.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 import { makeRedeemOpsAuditService } from './auditService.js';
 import { makeDedupeService } from './dedupeService.js';
