+++ backend/src/services/redeemOps/poolService.js
@@ -2,7 +2,7 @@ import { Op, QueryTypes } from 'sequelize';
 import {
   ProspectingPool, ProspectingPoolMember, PartnerOrganisation, sequelize,
 } from '../../models/index.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 import { makeClaimService } from './claimService.js';
