+++ backend/src/services/redeemOps/redemptionService.js
@@ -3,7 +3,7 @@ import {
   RewardEntitlement, Redemption, RedemptionEvent, RewardOffer, Activation,
   PartnerOrganisation, PartnerLocation, Prospect, User, sequelize,
 } from '../../models/index.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 import { makeInventoryService } from './inventoryService.js';
 import { makeRedeemOpsAuditService } from './auditService.js';
