+++ backend/src/services/redeemOps/rewardService.js
@@ -3,7 +3,7 @@ import {
   RewardOffer, RewardTermsVersion, RewardOfferLocation, RewardInventoryEvent,
   PartnerOrganisation, PartnerLocation, User, sequelize,
 } from '../../models/index.js';
-import { AppError } from '../../middleware/errorHandler.js';
+import { AppError } from '../../middleware/appError.js';
 import { logger } from '../../utils/logger.js';
 import { makeRedeemOpsAuditService } from './auditService.js';
 import { invalidateMarketplaceCache } from '../marketplaceCache.js';
